requirements.txt
django>=1.6.5,<1.7.0
celery>=3.1,<3.2
flower==0.7.0
django-frontend>=1.3.0,<1.4
Pillow
sorl-thumbnail
requests
